requirements.txt
boto3==1.20.24
pandas==1.2.5
python-dotenv==0.20.0
